pyproject.toml
[project]
name = "agents-intro"
version = "0.1.0"
description = "Intro to agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipython>=8.34.0",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.11",
    "python-dotenv>=1.0.1",
]
